+++ packages/dsw-data-seeder/dsw/data_seeder/cli.py
@@ -7,7 +7,7 @@ from dsw.config.parser import MissingConfigurationError
 
 from .config import SeederConfig, SeederConfigParser
 from .consts import PROG_NAME, VERSION, NULL_UUID
-from .logging import prepare_logging
+from .seeder import DataSeeder, SeedRecipe
 
 
 def validate_config(ctx, param, value: Optional[IO]):
@@ -33,16 +33,16 @@ def validate_config(ctx, param, value: Optional[IO]):
 
 @click.group(name=PROG_NAME)
 @click.version_option(version=VERSION)
-@click.option('-c', '--config', envvar='DSW_CONFIG',
+@click.option('-c', '--config', envvar='APPLICATION_CONFIG_PATH',
               required=False, callback=validate_config,
               type=click.File('r', encoding='utf-8'))
-@click.option('-w', '--workdir', envvar='SEEDER_WORKDIR',
+@click.option('-w', '--workdir', envvar='WORKDIR_PATH',
               type=click.Path(dir_okay=True, exists=True))
 @click.pass_context
 def cli(ctx: click.Context, config: SeederConfig, workdir: str):
     ctx.obj['cfg'] = config
     ctx.obj['workdir'] = pathlib.Path(workdir).absolute()
-    prepare_logging(cfg=config)
+    config.log.apply()
 
 
 @cli.command()
@@ -52,7 +52,6 @@ def run(ctx: click.Context, recipe: str):
     """Run worker that listens to persistent commands"""
     cfg = ctx.obj['cfg']
     workdir = ctx.obj['workdir']
-    from .seeder import DataSeeder
     seeder = DataSeeder(cfg=cfg, workdir=workdir)
     seeder.run(recipe)
 
@@ -65,7 +64,6 @@ def seed(ctx: click.Context, recipe: str, app_uuid: str):
     """Seed data in DSW directly"""
     cfg = ctx.obj['cfg']
     workdir = ctx.obj['workdir']
-    from .seeder import DataSeeder
     seeder = DataSeeder(cfg=cfg, workdir=workdir)
     seeder.seed(recipe_name=recipe, app_uuid=app_uuid)
 
@@ -75,7 +73,6 @@ def seed(ctx: click.Context, recipe: str, app_uuid: str):
 def list(ctx: click.Context):
     """List recipes for data seeding"""
     workdir = ctx.obj['workdir']
-    from .seeder import SeedRecipe
     recipes = SeedRecipe.load_from_dir(workdir)
     for recipe in recipes.values():
         click.echo(recipe)