+++ packages/dsw-data-seeder/dsw/data_seeder/config.py
@@ -1,23 +1,27 @@
 from dsw.config import DSWConfigParser
 from dsw.config.model import DatabaseConfig, S3Config, \
-    LoggingConfig, CloudConfig
+    LoggingConfig, SentryConfig, CloudConfig, GeneralConfig
 
 
 class SeederConfig:
 
     def __init__(self, db: DatabaseConfig, s3: S3Config, log: LoggingConfig,
-                 cloud: CloudConfig):
+                 sentry: SentryConfig, cloud: CloudConfig, general: GeneralConfig):
+        self.general = general
         self.db = db
         self.s3 = s3
         self.log = log
+        self.sentry = sentry
         self.cloud = cloud
 
     def __str__(self):
         return f'SeederConfig\n' \
                f'====================\n' \
+               f'{self.general}' \
                f'{self.db}' \
                f'{self.s3}' \
                f'{self.log}' \
+               f'{self.sentry}' \
                f'{self.cloud}' \
                f'====================\n'
 
@@ -30,5 +34,7 @@ class SeederConfigParser(DSWConfigParser):
             db=self.db,
             s3=self.s3,
             log=self.logging,
+            sentry=self.sentry,
             cloud=self.cloud,
+            general=self.general,
         )