+++ packages/dsw-mailer/dsw/mailer/context.py
@@ -2,7 +2,6 @@ import pathlib
 
 from typing import Optional, TYPE_CHECKING
 
-from .logging import LOGGER
 from .templates import TemplateRegistry
 
 if TYPE_CHECKING:
@@ -41,31 +40,23 @@ class _Context:
         self.templates = templates
         self.mode = mode
 
+    def update_trace_id(self, trace_id: str):
+        self.app.cfg.log.set_logging_extra('traceId', trace_id)
+        self.job.trace_id = trace_id
+
+    def reset_ids(self):
+        self.update_trace_id('-')
+
 
 class Context:
 
     _instance = None  # type: Optional[_Context]
-    logger = LOGGER
 
     @classmethod
     def get(cls) -> _Context:
         if cls._instance is None:
             raise ContextNotInitializedError()
         return cls._instance
-
-    @classmethod
-    def update_trace_id(cls, trace_id: str):
-        cls.logger.trace_id = trace_id
-        cls.get().job.trace_id = trace_id
-
-    @classmethod
-    def update_document_id(cls, document_id: str):
-        cls.logger.document_id = document_id
-
-    @classmethod
-    def reset(cls):
-        cls.update_trace_id('-')
-        cls.update_document_id('-')
 
     @classmethod
     def is_registry_mode(cls):