+++ packages/dsw-document-worker/CHANGELOG.md
@@ -8,6 +8,14 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
 
+## [3.23.0]
+
+### Changed
+
+- Improved command queue
+- Changed envvars for configuration
+- Enhanced error reporting
+
 ## [3.22.1]
 
 Released for version consistency with other DSW tools.
@@ -140,3 +148,4 @@ Released for version consistency with other DSW tools.
 [3.21.0]: /../../tree/v3.21.0
 [3.22.0]: /../../tree/v3.22.0
 [3.22.1]: /../../tree/v3.22.1
+[3.23.0]: /../../tree/v3.23.0